requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
greenlet==3.1.1
sqladmin==0.20.1
itsdangerous==2.2.0
httpx==0.28.1
beautifulsoup4==4.12.3
python-slugify==8.0.4
passlib[bcrypt]==1.7.4
